+++ client/src/containers/MapContainer/MapContainer.js
@@ -1,5 +1,3 @@
-/* global google */
-
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import * as mapActions from '../../actions/mapActions.js';