+++ client/src/reducers/map.js
@@ -1,9 +1,7 @@
-/* global google */
-
 import * as types from '../constants/actionTypes.js';
 
-export const HEATMAP_LAYER_NAME = 'HeatMap';
 export const CLUSTER_LAYER_NAME = 'Clusters';
+export const HEATMAP_LAYER_NAME = 'HeatMap';
 
 const initialState = {
   lonLat: [0, 0],
@@ -13,15 +11,15 @@ const initialState = {
   clickRadius: 250,
   layers: [
     {
-      label: HEATMAP_LAYER_NAME,
+      label: CLUSTER_LAYER_NAME,
       value: 1
     },
     {
-      label: CLUSTER_LAYER_NAME,
+      label: HEATMAP_LAYER_NAME,
       value: 2
     }
   ],
-  selectedLayer: 2
+  selectedLayer: 1
 };
 
 export default function map(state = initialState, action) {